requirements.txt
mistralai
dotenv
voyageai
pinecone
google.generativeai